Fix DrawFirstRow UNIT_nSM dataToShow
Now conversion can benefit from the internal FPU.
((float) counter / 3.54f note that f suffix is to
tell the compiler 3.54 is a float.

+++ Src/lcddraw.c
@@ -33,7 +33,7 @@ void DrawFirstRow(const uint64_t secondsSinceStartup, const uint64_t potvalue, c
 		break;
 	case UNIT_nSM:
 		StampaStringaSuLCD(ROW_SHORT_COUNTER, N_COLONNE - 5 - digits, "nS/m~"); // TODO: fix cast
-		dataToShow = (uint64_t) ((double) counter / 3.54);
+		dataToShow = (uint64_t) ((float) counter / 3.54f); // f suffix clarifies that 3.54 is to be treated as float
 		break;
 	}
